Issue SC-1951 feat: Code refactor - Moved the common logic to util file
