+++ data-products/src/main/scala/org/sunbird/analytics/job/report/AssessmentMetricsJobV2.scala
@@ -60,23 +60,6 @@ object AssessmentMetricsJobV2 extends optional.Application with IJob with BaseRe
     fc.closeContext()
   }
 
-//  /**
-//   * Generic method used to load data by passing configurations
-//   *
-//   * @param spark    - Spark Sessions
-//   * @param settings - Cassandra/Redis configs
-//   * @param url      - Cassandra/Redis url
-//   * @return
-//   */
-//  def loadData(spark: SparkSession, settings: Map[String, String], url: String, schema: StructType): DataFrame = {
-//    if (schema.nonEmpty) {
-//      spark.read.schema(schema).format(url).options(settings).load()
-//    }
-//    else {
-//      spark.read.format(url).options(settings).load()
-//    }
-//  }
-
   def prepareReport(spark: SparkSession, loadData: (SparkSession, Map[String, String], String, Option[StructType], Option[Seq[String]]) => DataFrame, config: JobConfig, batchList: List[String])(implicit fc: FrameworkContext): Unit = {
 
     implicit val sparkSession: SparkSession = spark
@@ -84,7 +67,7 @@ object AssessmentMetricsJobV2 extends optional.Application with IJob with BaseRe
     import sqlContext.implicits._
     val batches = CourseUtils.getActiveBatches(loadData, batchList, sunbirdCoursesKeyspace)
     val userData = CommonUtil.time({
-      CourseUtils.recordTime(getUserData(spark, loadData), "Time taken to get generate the userData- ")
+      CourseUtils.recordTime(CourseUtils.getUserData(spark, loadData), "Time taken to get generate the userData- ")
     })
     val modelParams = config.modelParams.get
     val contentFilters = modelParams.getOrElse("contentFilters",Map()).asInstanceOf[Map[String,AnyRef]]
@@ -129,19 +112,10 @@ object AssessmentMetricsJobV2 extends optional.Application with IJob with BaseRe
     userData._2.unpersist(true)
   }
 
-  def getUserData(spark: SparkSession, loadData: (SparkSession, Map[String, String], String, Option[StructType], Option[Seq[String]]) => DataFrame): DataFrame = {
-    val schema = Encoders.product[UserData].schema
-    loadData(spark, Map("table" -> "user", "infer.schema" -> "true", "key.column" -> "userid"), "org.apache.spark.sql.redis", Some(schema), Some(Seq("firstname", "lastname")))
-      .withColumn("username", concat_ws(" ", col("firstname"), col("lastname"))).persist(StorageLevel.MEMORY_ONLY)
-  }
 
-  def getUserEnrollmentDF(loadData: (SparkSession, Map[String, String], String, Option[StructType], Option[Seq[String]]) => DataFrame)(implicit spark: SparkSession): DataFrame = {
-    loadData(spark, Map("table" -> "user_enrolments", "keyspace" -> sunbirdCoursesKeyspace), cassandraUrl, Some(new StructType()), Some(Seq("batchid","userid","courseid","active","completionpercentage","enrolleddate","completedon")))
-      .filter(lower(col("active")).equalTo("true"))
-  }
 
   def getAssessmentProfileDF(loadData: (SparkSession, Map[String, String], String, Option[StructType], Option[Seq[String]]) => DataFrame)(implicit spark: SparkSession): DataFrame = {
-    val userEnrolmentDF = getUserEnrollmentDF(loadData).persist(StorageLevel.MEMORY_ONLY)
+    val userEnrolmentDF = CourseUtils.getUserEnrollmentDF(loadData, keyspace = sunbirdCoursesKeyspace).persist(StorageLevel.MEMORY_ONLY)
     val assessmentProfileDF = loadData(spark, Map("table" -> "assessment_aggregator", "keyspace" -> sunbirdCoursesKeyspace), cassandraUrl, Some(new StructType()), Some(Seq("course_id","batch_id","user_id","content_id","total_max_score","total_score","grand_total")))
     val assessmentDF = getAssessmentData(assessmentProfileDF)
